+++ scripts/deploy-addons.sh
@@ -11,14 +11,16 @@ case "${ENVIRONMENT}" in
     ;;
 esac
 
-helm repo add ingress-nginx https://kubernetes.github.io/ingress-nginx
+kubectl apply -f https://github.com/kubernetes-sigs/gateway-api/releases/download/v1.4.1/standard-install.yaml
+
+helm install ngf oci://ghcr.io/nginx/charts/nginx-gateway-fabric \
+  --create-namespace \
+  -n nginx-gateway
+
 helm repo add jetstack https://charts.jetstack.io
 helm repo update
 
-helm upgrade --install ingress-nginx ingress-nginx/ingress-nginx \
-  -f "infra/helm/ingress-nginx/values-${ENVIRONMENT}.yaml" \
-  -n ingress-nginx --create-namespace
-
 helm upgrade --install cert-manager jetstack/cert-manager \
   -f "infra/helm/cert-manager/values-${ENVIRONMENT}.yaml" \
-  -n cert-manager --create-namespace
+  -n cert-manager --create-namespace \
+  --set installCRDs=true